Change price variable to double, and add space

+++ itc115/itc115-assignment10/src/Ticket.java
@@ -1,10 +1,12 @@
 
 public abstract class Ticket {
 	int number;
-	int price;
+	double price;
+	
 	Ticket(int number){
 		this.number = number;
 	}
+	
 	public abstract double getPrice();
 	public abstract String toString();
 }
